+++ src/conversion/convert_text.ts
@@ -15,11 +15,11 @@ import {
 	TFile,
 	Vault,
 } from "obsidian";
-import { getDataviewPath } from "./filePath";
+import { getDataviewPath } from "./file_path";
 import { getAPI, Link } from "obsidian-dataview";
 import { noticeLog } from "../src/utils";
 import { convertLinkCitation, convertWikilinks } from "./links";
-import findAndReplaceText from "./findAndReplaceText";
+import findAndReplaceText from "./find_and_replace_text";
 import GithubPublisher from "../main";
 import i18next from "i18next";
